third_party/WebKit/LayoutTests/imported/wpt/html/webappapis/scripting/processing-model-2/compile-error-in-setTimeout-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected end of input
This is a testharness.js-based test.
FAIL window.onerror - compile error in setTimeout assert_equals: second arg expected "file:///.../html/webappapis/scripting/processing-model-2/compile-error-in-setTimeout.html" but got ""
PASS window.onerror - compile error in setTimeout (column) 
Harness: the test ran to completion.
